+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 This file is used to list changes made in each version of dop_php.
 
+## 3.0.4 (Unreleased)
+
+* do not add php7.0-oauth by default on debian systems (as dotdeb don't provide the php7.0-oauth package) (sheuser)
 
 ## 3.0.3 (Unreleased)